requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
